Renamed .h to .hpp for files with included definition and no cpp
